+++ light-curve/light_curve/embed/__init__.py
@@ -1,5 +1,6 @@
 from .astra_clr import AstraCLR
 from .astromer import Astromer1, Astromer2
+from .atat import ATAT
 from .atcat import ATCAT
 from .model import EmbeddingSession, SingleBandModel
 from .reduction import (
@@ -16,6 +17,7 @@ __all__ = [
     "AstraCLR",
     "Astromer1",
     "Astromer2",
+    "ATAT",
     "ATCAT",
     "Beginning",
     "EmbeddingSession",